Drop stray semicolon and useless cast
Vault::SetErrorCallback carried a trailing semicolon after its in-class
definition, unlike the equivalent AesGcm::SetErrorCallback in either
project. The vault file test cast an expression to size_t that the usual
arithmetic conversions had already made size_t.

Add -Wextra-semi to BASE_WARNINGS and the GCC-only -Wuseless-cast to
STRICT_WARNINGS so both stay fixed.

+++ Projects/PasswordManager/src/tests/vault_file_test.cpp
@@ -416,7 +416,7 @@ TEST_F(VaultFileTest, SaveWritesFreshIV) {
   std::vector<uint8_t> second = ReadFile(path_);
 
   ASSERT_EQ(first.size(), second.size());
-  ASSERT_GT(first.size(), static_cast<size_t>(kMagicSize + kSaltSize + kIVSize + kTagSize));
+  ASSERT_GT(first.size(), kMagicSize + kSaltSize + kIVSize + kTagSize);
 
   const size_t salt_off = kMagicSize;
   const size_t iv_off = kMagicSize + kSaltSize;
